Enum taskqueue::LoopResult

pub enum LoopResult<R: Send + 'static> {
    Continue,
    Done(R),
}

Helper struct for Queue::loop_while

Inside your loop body return either Continue to let your loop continue running or return Done with the return value, that shall be observable through the FutureGuard loop_while returns.

Variants

Continue

Let loop continue

Done

Abort loop and return value