Trait taskqueue::extensions::iterator::ConcurrentMap

pub trait ConcurrentMap: Iterator + Send + Sized {
    fn concurrent_map<B, F>(self, f: F) -> IntoIter<B> where F: Fn(Self::Item) -> B + Send + Sync, B: Send + 'static;
}

The ConcurrentMap Trait adds a map like function to iterators that utilizes TaskQueue to execute it in parallel.

Required Methods

fn concurrent_map<B, F>(self, f: F) -> IntoIter<B> where F: Fn(Self::Item) -> B + Send + Sync, B: Send + 'static

map like function utilizing a ConcurrentQueue internally to execute it parallely returning a Group which also implements the Iterator Trait and holds the results in order.

Example

let squared: Vec<i32> = (0..20).concurrent_map(|x| x*x).collect();