Module taskqueue::io

IO module providing a queue-compatible mostly std-like replacement for native blocking IO. Provided through mioco and mio. Other types implementing mio's Evented Trait might be used aswell through io::mio::MioAdapter.

With queues executing on a thread pool, normal blocking IO is not working anymore. Anything blocking may block multiple queues at once causing a deadlock or starvation.

In cases where blocking IO cannot be avoided, a solution is provided through SerialQueue::new_native().

Modules

mail
Mailboxes

sync
Useful synchronization primitives

tcp
TCP IO

timer
Timers

udp
UDP IO

unix
Unix sockets IO

Structs

MioAdapter
Adapt raw mio type to mioco Evented requirements.