Struct taskqueue::BoundSerialQueue

pub struct BoundSerialQueue<'queue, T: 'static> {
    // some fields omitted
}

A bound SerialQueue holding a queue-bound variable

Create a BoundSerialQueue using SerialQueue::with. BoundSerialQueue's hold variables that may be used through tasks executed on this queue though scoped_with, sync_with foreach_with or loop_while_with.

async_with cannot be provided, as the bound variable is dropped, when the BoundSerialQueue gets dropped.

Internally BoundSerialQueue refer to the same queue, they were created from. Multiple BoundSerialQueue's may exist for one queue at once.

Methods

impl<'queue, T: 'static> BoundSerialQueue<'queue, T>

fn scoped_with<R, F>(&'queue self, operation: F) -> FutureGuard<R> where R: Send + 'static, F: FnOnce(&'queue mut T) -> R + Send + 'queue

Like Queue::scoped but provides a mutable reference to the bound variable

Safety

The same rules as for Queue::scoped to apply.

Example

let bound = queue.with(|| { "Hello".to_string() });
let name = "Stack".to_string();
bound.scoped_with(|message| { println!("{} {}!", message, name) });

fn sync_with<R, F>(&'queue self, operation: F) -> R where R: Send + 'static, F: FnOnce(&'queue mut T) -> R + Send + 'queue

Like Queue::sync but provides a mutable reference to the bound variable

Example

let bound = queue.with(|| { "a bound queue".to_string() });
bound.sync_with(|name| { println!("Hello {}", name) });

fn foreach_with<B, R, I, F>(&'queue self, iter: I, operation: F) -> Group<R> where B: Send, R: Send + 'queue, I: Iterator<Item=B> + Send, F: Fn(&'queue T, B) -> R + Send + Sync + 'queue

Like Queue::foreach but provides a mutable reference to the bound variable

Example

let bound = queue.with(|| 2);
let doubled: Vec<i32> = bound.foreach_with((0..20), |factor, x| x*factor).wait().collect();

fn loop_while_with<R, F>(&'queue self, operation: F) -> FutureGuard<R> where F: Fn(&'queue mut T) -> LoopResult<R> + Send + Sync + 'queue, R: Send + 'static

Like Queue::loop_while but provides a mutable reference to the bound variable

Example

use std::mem;

let queue = SerialQueue::new();
let bound = queue.with(|| (15, 25));

let greatest_common_divisor = bound.loop_while_with(|tuple| {
    let x = tuple.0;
    let y = tuple.1;

    if y == 0 {
        LoopResult::Done(x.clone())
    } else {
        let remainder = x % y;
        let mut new_tuple = (y, remainder);
        mem::swap(tuple, &mut new_tuple);
        LoopResult::Continue
    }
}).get();