Struct taskqueue::ConcurrentQueue

pub struct ConcurrentQueue {
    // some fields omitted
}

Queue executing tasks in parallel

Properties

executes a new task as soon as possible on the next thread

tasks may overlap

tasks may execute on totally different threads

safety against deadlocks from recursive queueing (see SerialQueue example)

This disallows type binding but enables very easy parallization where appropriate.

Example

let queue = ConcurrentQueue::new();

let mut result = Vec::with_capacity(4096);
for x in 0..4096 {
    result.push(queue.async(move || x*x));
}
let result: Vec<i32> = result.into_iter().map(|x| x.get()).collect();

Methods

impl ConcurrentQueue

fn new() -> ConcurrentQueue

Create a new ConcurrentQueue and executing tasks on the global thread pool