Struct taskqueue::Future

pub struct Future<R: Send + 'static> {
    // some fields omitted
}

Future holding the variable of a maybe to be executed task

Methods

impl<R: Send + 'static> Future<R>

fn get(self) -> R

Blocks the current Queue until the value can be recieved. This may happen instantly or take a long time, completely depending on the task and the amount of work the corresponding queue has to face.