Struct taskqueue::Group

pub struct Group<R: Send + 'static> {
    // some fields omitted
}

Group provides a way to synchronize multiple Tasks

Using Group you can block on multiple tasks at once or handle them in order across multiple queues.

Methods

impl<R: Send + 'static> Group<R>

fn new() -> Group<R>

Create a new Group

fn async<T, F>(&mut self, queue: &T, task: F) where F: FnOnce() -> R + Send + 'static, T: Queue

Add a task executed on the passed queue to the Group. Behaves otherwise like Queue::async.

Arguments

queue - The Queue to execute upon task - The Task to be executed

Example

let queue_one = SerialQueue::new();
let queue_two = SerialQueue::new();
let mut group = Group::new();
group.async(&queue_one, || {
    // some heavy computation
});
group.async(&queue_two, || {
    // another heavy computation
});
group.wait();

fn scoped<'queue, T, F>(&mut self, queue: &'queue T, task: F) where F: FnOnce() -> R + Send + 'queue, T: Queue

Add a task executed on the passed queue to the Group. Behaves otherwise like Queue::scoped, so it may capture stack variables.

The Guard is not lost when dropping the Group, but is instead also dropped blocking till the end of the task execution.

Safety

Same rules as for Queue::scoped FutureGuard return value do apply to the Group, when this function was used.

fn wait(self) -> IntoIter<R>

Blocks the current Queue until all tasks finished execution and returns it as Iterator