Struct taskqueue::SerialQueue

pub struct SerialQueue {
    // some fields omitted
}

Queue executing Tasks serially, non-overlapping in queued Order

Properties

executes tasks in serial order

no tasks may overlap

they never change their native background thread (but may share it)

the tasks are executed in order of queuing

safety against deadlocks from recursive queueing (see example)

Through these guarantees SerialQueues may bound to a type that is not Send or Sync and provide easy thread safe access to this critcal resource. Such a SerialQueue is called BoundSerialQueue.

Example

init_main(|main| {
    let thread_one = SerialQueue::new();
    let thread_two = SerialQueue::new();

    let future_one = thread_one.async(|| {
        42
    });
    let future_two = thread_two.async(|| {
        96
    });

    println!("Although this is happening in main,");
    main.async(|| {
        println!("this task is running before...");
    });
    main.sync(|| {
        println!("...this task and...");
        assert_eq!(future_one.get() + future_two.get(), 138);
    });
    println!("...this is running last");
});

Methods

impl SerialQueue

fn new() -> SerialQueue

Create a new SerialQueue and assign it to the global thread pool

fn new_native() -> SerialQueue

Create a new SerialQueue and assign it solely to a newly created OS Thread

A SerialQueue created through this method will spawn a new native OS Thread and the queue will be the only utilizing it. The thread will be destructed, when the queue is dropped.

The purpose of this constructor is to provide a way to use blocking IO with TaskQueue. The use of this method however is discouraged, as the new thread may influence the scheduler negatively and evented IO, where possible performs a lot better in combination with the TaskQueue library

fn with<'queue, R: 'static, F>(&'queue self, constructor: F) -> BoundSerialQueue<'queue, R> where F: FnOnce() -> R + Send

Bind this queue to a variable

This function allows to create a BoundSerialQueue. Its purpose is to bind variables to a queue, so they can be used by the tasks submitted.

Example

let x = 5;
let bound = queue.with(move || x);
bound.scoped_with(|x| println!("{}", x));
// x gets dropped inside the queues thread before the queue gets dropped

You can create multiple bindings at once. Through tuples you may bind multiple variables at once.

It is even possible to move the creation of the bound variable into the queue by creating it inside the passed constructor, which is then executed on the queue. And because SerialQueues never change their underlying OS Thread, this allows to use variables that are not Send and Sync in a thread-safe but shared way.

Example

let bound = queue.with(|| {
    let raw_ptr = my_ffi_function();
    raw_ptr
});
bound.scoped_with(|raw_ptr| println!("{}", raw_ptr.is_null()));
// raw_ptr gets dropped inside the queues thread.
// This way raw_ptr is never moved between threads.