Trait taskqueue::Queue

pub trait Queue: Send + Sync + Sized + Debug {
    fn async<R, F>(&self, task: F) -> Future<R> where R: Send + 'static, F: FnOnce() -> R + Send + 'static;

    fn scoped<'queue, R, F>(&'queue self, task: F) -> FutureGuard<R> where R: Send + 'static, F: FnOnce() -> R + Send + 'queue { ... }
    fn sync<'queue, R, F>(&'queue self, task: F) -> R where R: Send + 'static, F: FnOnce() -> R + Send + 'queue { ... }
    fn foreach<'queue, B, R, I, F>(&'queue self, iter: I, task: F) -> Group<R> where B: Send, R: Send + 'queue, I: Iterator<Item=B> + Send, F: Fn(B) -> R + Send + Sync + 'queue { ... }
    fn loop_while<'queue, R, F>(&'queue self, task: F) -> FutureGuard<R> where F: Fn() -> LoopResult<R> + Send + Sync + 'queue, R: Send + 'static { ... }
}

Basic Queue Trait defining all possible operations on Queues

Queues implement this, so you may use a unified Interface for all Queues.

Required Methods

fn async<R, F>(&self, task: F) -> Future<R> where R: Send + 'static, F: FnOnce() -> R + Send + 'static

Runs a task asynchrously on this queue

async may not capture values on the current stack, as the closure may live longer then the current stack. For example, if you drop the Future and the Queue directly afterwards. Your task will still be executed in that case.

Arguments

task - Task to be run by the queue

Example

queue.async(|| { println!("Hello Queue!") });

Note that this limitation does not mean you cannot move variables, this is totally fine:

let name = "Welt".to_string();
queue.async(move || { println!("Hello {}!", name) });

Provided Methods

fn scoped<'queue, R, F>(&'queue self, task: F) -> FutureGuard<R> where R: Send + 'static, F: FnOnce() -> R + Send + 'queue

Runs a task asynchrously on this queue allowing capture of stack variables

scoped may capture stack variables. To ensure the task may not outlive the current stack scope the returned FutureGuard blocks on drop, waiting for the task to finish executing. To avoid this behavior use async.

Safety

Note that it is unsafe to leak the returned FutureGuard. Neither mem::forget nor cyclic RA II structures should be used in combination with FutureGuard as a leak may result in a dangeling pointer (read crash). This is the same reason thread::scoped was removed from the standard library. Read more here 24292.

Arguments

task - Task to be run on this queue

Example

let name = "Stack".to_string();
queue.scoped(|| { println!("Hello {}!", name) }); //note the missing `move`
// join to the task occurs here

fn sync<'queue, R, F>(&'queue self, task: F) -> R where R: Send + 'static, F: FnOnce() -> R + Send + 'queue

Runs a task synchronously on this queue

sync returns after the task was executed blocking the current queue. As a result capturing of stack variables is totally safe with this function and no Future is required.

Arguments

task - Task to be run on this queue

Example

println!("Hello {}", queue.sync(|| { "from another queue".to_string() }));

fn foreach<'queue, B, R, I, F>(&'queue self, iter: I, task: F) -> Group<R> where B: Send, R: Send + 'queue, I: Iterator<Item=B> + Send, F: Fn(B) -> R + Send + Sync + 'queue

Runs a task once for each iterator elem on this queue

foreach queues one task multiple times for each iterator element of the iterator passed consuming it. It returns a Group containing the FutureGuards of all executions. Note that Group also implements the Itertor trait.

Arguments

iter - Iterator to be consumed task - Task to be run for iter on this queue

Example

let squared: Vec<i32> = queue.foreach((0..20), |x| x*x).wait().collect();

fn loop_while<'queue, R, F>(&'queue self, task: F) -> FutureGuard<R> where F: Fn() -> LoopResult<R> + Send + Sync + 'queue, R: Send + 'static

Loops a task on this queue while a condition is hold

loop_while queues a task again and again until it returns Done. Note that this is perferred over:

queue.async(|| {
    loop { //this is bad
         /* your task */
    }
});

because loop_while allows other queued tasks to run in the meantime, because the task is queued again and again, while the code above runs as one long task.

Note that this is especially useful in combination with SerialQueue::init_main to create a main thread application run loop.

Arguments

task - Task to be looped on this queue

Example

let program = Program;

queue.loop_while(move || {
   match program.did_exit() {
       Some(exit_code) => LoopResult::Done(exit_code),
       None => {
           program.tick();
           LoopResult::Continue
      },
   }
}).get();